+++ src/Engines/Ruleset.php
@@ -9,7 +9,6 @@ use Kettasoft\Filterable\Traits\FieldNormalizer;
 use Kettasoft\Filterable\Engines\Foundation\Engine;
 use Kettasoft\Filterable\Engines\Foundation\PayloadApplier;
 use Kettasoft\Filterable\Engines\Foundation\PayloadFactory;
-use Kettasoft\Filterable\Engines\Foundation\Appliers\Applier;
 use Kettasoft\Filterable\Engines\Foundation\Parsers\Dissector;
 
 class Ruleset extends Engine
@@ -43,7 +42,7 @@ class Ruleset extends Engine
           new Payload($field, $dissector->operator, $this->sanitizeValue($field, $dissector->value), $dissector->value)
         );
 
-        Applier::apply(new PayloadApplier($payload), $builder);
+        (new PayloadApplier($payload))->apply($builder);
 
         return $this->commit($field, $payload);
       });